+++ tid_project/src/components/layout/DashboardLayout.js
@@ -37,7 +37,7 @@ const DashboardLayout = () => {
               <Link to="/frame1" component={StyledNavLink}>Frame 1</Link>
               <Link to="/frame2" component={StyledNavLink}>Frame 2</Link>
               <Link to="/frame3" component={StyledNavLink}>Frame 3</Link>
-              <Link to="/frame4" component={StyledNavLink}>Frame 4</Link>
+              <Link to="/Threads" component={StyledNavLink}>Threads</Link>
               <StyledNavLink as="button" onClick={handleLogout}>
                 Logout
               </StyledNavLink>
