+++ tid_project/src/pages/CreateContent/CreateContentPage.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { ThreadProvider } from "../Frame4/ThreadContext"
+import { ThreadProvider } from "../Threads/ThreadContext"
 import { useNavigate } from 'react-router-dom';
 import { 
   ContentPageContainer,
