+++ tid_project/src/pages/Threads/ThreadCard.js
